Extract tool logic

+++ app_md/base_app.py
@@ -4,7 +4,7 @@ import traceback
 import qdarkstyle
 
 
-from PyQt5.QtWidgets import QAction, QApplication, QFileDialog, QLabel, QLineEdit, QMainWindow, QMessageBox, QPlainTextEdit, QPushButton, QSplashScreen, QVBoxLayout, QWidget
+from PyQt5.QtWidgets import QAction, QApplication, QCheckBox, QFileDialog, QLabel, QLineEdit, QMainWindow, QMessageBox, QPlainTextEdit, QPushButton, QSplashScreen, QVBoxLayout, QWidget
 from PyQt5.QtGui import QFont, QGuiApplication, QIcon, QKeySequence, QPixmap
 from PyQt5.QtCore import QFile, Qt, QTimer, QThreadPool
 from pathlib import Path
@@ -45,7 +45,7 @@ class BaseApp:
         # Cargar imagen del splash
         splash_pix = QPixmap(str(Path(__file__).resolve().parent / "images" / "splash.png"))
         self.splash = QSplashScreen(splash_pix, Qt.WindowStaysOnTopHint)
-        self.splash.showMessage("Cargando...", Qt.AlignBottom | Qt.AlignCenter, Qt.white)
+        self.splash.showMessage("Loading...", Qt.AlignBottom | Qt.AlignCenter, Qt.white)
         self.splash.show()
 
         # Esperar y luego lanzar la ventana principal
@@ -133,7 +133,7 @@ class BaseApp:
 
         # Si se obtuvo un archivo valido (arrastrado o desde dialogo)
         if file_path:
-            self.path_iso = file_path
+            self.path_iso = Path(file_path)
             self.window.label.setPlainText(file_path)
             self.window.success_dialog(vaule=["File loaded"])
 
@@ -163,7 +163,7 @@ class MainWindow(QMainWindow):
         self.thread_pool = QThreadPool()
 
         self.setWindowTitle(f"Editor indice tag team - v:{self.version}")
-        self.setFixedSize(550, 255)
+        self.setFixedSize(550, 280)
         self.setWindowIcon(QIcon(str(self.icon_path)))
 
         #acepta drop
@@ -194,6 +194,12 @@ class MainWindow(QMainWindow):
         self.label.setReadOnly(True)
         self.label.setLineWrapMode(QPlainTextEdit.NoWrap)
 
+        # Checkboxes adicionales
+        self.ischeckbox_wavs = False
+        self.checkbox_wavs = QCheckBox("Process WAV files to AT3", self)
+        self.checkbox_wavs.stateChanged.connect(self.on_state_checbox_wavs)
+        # self.checkbox_opcion2 = QCheckBox("Usar compresión avanzada", self)
+
         layout = QVBoxLayout()
         layout.addWidget(self.edit_lb_pack)
         layout.addWidget(self.edit_lbl_files)
@@ -201,10 +207,15 @@ class MainWindow(QMainWindow):
         layout.addWidget(self.boton_extiso)
         layout.addWidget(self.boton_compiso)
         layout.addWidget(self.label)
+        layout.addWidget(self.checkbox_wavs)
+        # layout.addWidget(self.checkbox_opcion2)
 
         # Asignar el layout al widget central
         central_widget.setLayout(layout)
 
+    def on_state_checbox_wavs(self, state):
+        self.ischeckbox_wavs = (state == Qt.Checked)
+        print(self.ischeckbox_wavs)
 
     def dragEnterEvent(self, event):
         #verifica si lo arrastrado son archivos
@@ -222,9 +233,6 @@ class MainWindow(QMainWindow):
             self.manejar_error("Only one file is allowed.")
             return  # Ignora si hay mas de uno
 
-        #confirmacion para cargar el path
-        if self.question_dialog("are you sure you want open iso file?") == QMessageBox.Cancel:
-            return
 
         filepath = urls[0].toLocalFile()
         #confirmacion para abrir el iso
@@ -234,7 +242,7 @@ class MainWindow(QMainWindow):
             self.contenedor.open_iso(file_path=filepath)
 
     def closeEvent(self, event):
-        reply = self.question_dialog(content="Are you sure you want to close the application?", title="Confirm exit")
+        reply = self.question_dialog(content="Are you sure you want to close the Editor application?", title="Confirm exit")
 
         if reply == QMessageBox.Ok:
             event.accept()
@@ -284,6 +292,8 @@ class MainWindow(QMainWindow):
         def delete_content_folder(path_folder):
             for elemento in os.listdir(path_folder):
                 ruta_elemento = os.path.join(path_folder, elemento)
+                if Path(ruta_elemento).suffix in ".json":
+                    continue
                 if os.path.isfile(ruta_elemento) or os.path.islink(ruta_elemento):
                     os.remove(ruta_elemento)
                 elif os.path.isdir(ruta_elemento):
@@ -355,8 +365,8 @@ class MainWindow(QMainWindow):
             self.manejar_error(f"The folder \"{self.new_folder.name}\" does not exist in the file path.")
             return
 
-        if QFile.exists(self.contenedor.path_iso+".compress"):
-            respuesta = self.question_dialog("The iso.compress exists; its file will be deleted.")
+        if QFile.exists(str(self.contenedor.path_iso.parent / f"compress_{self.contenedor.path_iso.name}")):
+            respuesta = self.question_dialog("The iso compress exists; its file will be deleted.")
             # respuesta = QMessageBox.question(
             #                 self,
             #                 "Warning!",
@@ -387,18 +397,23 @@ class MainWindow(QMainWindow):
 
     def manejar_error(self, error_msg):
         #mostrar una ventana con el error
+        self.contenedor.extract_w.setEnabled(True)
         self.setEnabled(True)
+
         QApplication.restoreOverrideCursor()
         ErrorDialog(error_msg, self.icon_path).exec_()
 
     def success_dialog(self, vaule, title:str="Success"):
+        self.contenedor.extract_w.setEnabled(True)
         self.setEnabled(True)
+
         QApplication.restoreOverrideCursor()
+
         if 'href' not in vaule[0]:
             QMessageBox.information(self, title, vaule[0])
             return
 
-        self.secundaria = Open_folder_link(parent_font=self.font(), parent_icon=self.windowIcon(), messag=vaule[0], direc=self.new_folder)
+        self.secundaria = Open_folder_link(parent_font=self.font(), parent_icon=self.windowIcon(), messag=vaule[0], direc=self.new_folder if not isinstance(vaule[-1], Path) else vaule[-1])
         self.secundaria.exec_()
 
     def question_dialog(self, content, title:str="Warning!"):
